+++ react-components/src/components/Divider/model/index.ts
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export interface DividerProps {
+  // Направление
+  type?: 'horizontal' | 'vertical';
+
+  // Длина
+  length?: React.CSSProperties['height'];
+}