+++ react-components/src/components/Stack/index.tsx
@@ -1,10 +1,11 @@
 import React, { Children, FC } from 'react';
 import classNames from 'classnames';
 
+import { Divider } from '../Divider';
 import { StackProps } from './model';
+import { getGaps } from './utils/getGaps';
 
 import s from './style.module.scss';
-import { getGaps } from './utils/getGaps';
 
 export const Stack: FC<StackProps> = ({
   direction = 'row',
@@ -41,9 +42,16 @@ export const Stack: FC<StackProps> = ({
       }}
     >
       {Children.map(children, (child, index) => {
+        const dividerType = isColumnDirection ? 'horizontal' : 'vertical';
         const dividerElement =
           Children.count(children) !== index + 1 ? (
-            <div className={s.divider} />
+            <div className={s.divider}>
+              {isBoolean(divider) ? (
+                <Divider length="var(--spacing-16)" type={dividerType} />
+              ) : (
+                divider
+              )}
+            </div>
           ) : null;
 
         if (!child) return null;
@@ -59,3 +67,7 @@ export const Stack: FC<StackProps> = ({
     </div>
   );
 };
+
+function isBoolean(value: StackProps['divider']): value is boolean {
+  return typeof value === 'boolean';
+}