+++ react-components/src/index.ts
@@ -41,3 +41,5 @@ export * from './components/List';
 export * from './components/List/model/public-api';
 
 export * from './components/Slider';
+
+export * from './components/Divider';